chore(telemetry): more cleanup: remove apis.telemetry
# What does this PR do?


## Test Plan
